Add out-of-line copy ctor to Maps::Iterator.
Chrome's style checking plugin complains that this class needs an explicit
out-of-line copy constructor.

So far, the style plugin has not warned about nested classes because of a bug,
but we have fixed that bug and are now cleaning up all the cases where it warns.

Committed on behalf of Hans Wennborg (hans@chromium.org).

BUG=http://code.google.com/p/chromium/issues/detail?id=139346

Review URL: https://chromiumcodereview.appspot.com/10824229